+++ client/src/pages/contextualActions/ContextualActionsLauncher.tsx
@@ -1,60 +1,23 @@
-import { createContext, ReactNode, useContext, useState } from "react";
 import * as React from "react";
+import { useContext } from "react";
 import styled from "styled-components";
 import { useFlag } from "../../common/Hooks";
 import { If } from "../../common/If";
-import { DropDown } from "../../common/searchBox/dropDown/DropDown";
-import { IDropDownItem } from "../../common/searchBox/dropDown/IDropDownItem";
-
-export interface IContextualAction {
-  key: string;
-  label: ReactNode;
-  onClick: () => void;
-}
-
-export const ContextualActionsContext = createContext<ConActConObj>(null);
-
-export class ConActConObj {
-  public constructor(
-    public actions: IContextualAction[] = [],
-    private setActions: (actions: IContextualAction[]) => void
-  ) {}
-
-  public addAction = (action: IContextualAction): void => {
-    if (!this.containsAction(action)) {
-      this.actions = [...this.actions, action];
-      this.setActions(this.actions);
-    }
-  };
-
-  public removeAction = (actionKey: string): void => {
-    this.actions = this.actions.filter(a => a.key !== actionKey);
-    this.setActions(this.actions);
-  };
-
-  private containsAction = (action: IContextualAction): boolean => {
-    return this.actions.filter(a => a.key === action.key).length > 0;
-  };
-}
-
-export const ContextualActionsProvider = (props: { children: ReactNode }) => {
-  const [actions, setActions] = useState<IContextualAction[]>([]);
-
-  return (
-    <ContextualActionsContext.Provider value={new ConActConObj(actions, setActions)}>
-      {props.children}
-    </ContextualActionsContext.Provider>
-  );
-};
+import { ContextualActionsContext } from "./ContextualActionsContext";
+import { ContextualActionsPanel } from "./ContextualActionsPanel";
 
 export const ContextualActionsLauncher = () => {
   const [isOpen, setIsOpen] = useFlag(false);
+  const contextualActionsContext = useContext(ContextualActionsContext);
 
   return (
     <>
-      <Container onClick={setIsOpen as any}>@</Container>
       <If
-        value={isOpen}
+        value={contextualActionsContext.actions}
+        render={() => <Container onClick={setIsOpen as any}>@</Container>}
+      />
+      <If
+        value={isOpen && contextualActionsContext.actions}
         render={() => <ContextualActionsPanel closePanel={() => setIsOpen(false)} />}
       />
     </>
@@ -71,37 +34,4 @@ const Container = styled.div`
   color: ${p => p.theme.colors.accentContrast};
   font-weight: ${p => p.theme.font.weight.bold};
   cursor: pointer;
-`;
-
-export const ContextualActionsPanel = (props: { closePanel: () => void }) => {
-  const contextualActionsContext = useContext(ContextualActionsContext);
-
-  return (
-    <Panel>
-      <DropDown
-        groups={[
-          {
-            onSelectItem: (item: IDropDownItem<IContextualAction>) => {
-              if (item && item.item.onClick) {
-                item.item.onClick();
-              }
-            },
-            items: contextualActionsContext.actions.map(a => ({
-              key: a.key,
-              label: a.label,
-              item: a
-            }))
-          }
-        ]}
-        onClose={props.closePanel}
-      />
-    </Panel>
-  );
-};
-
-const Panel = styled.div`
-  width: 300px;
-  position: absolute;
-  right: 50px;
-  top: 0;
 `;
